Trying to get profile_guider running with the new guiderImages repo.
There were a bunch of testing API changes that profile_guider didn't keep up
with and the files were moved to a separate repo.

It's not quite working (gave me an exception when writing the FITS file), but
it should be close.

Also did pep8 cleanups.

(the goal here was a quick test of PyGuide performance for LSST)